requirements.txt
Pillow
qrcode[pil]
